pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "phone-farm"
version = "0.1.0"
description = "Android phone farm for automated app testing"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1",
    "aiosqlite>=0.20",
    "cryptography>=42.0",
    "Appium-Python-Client>=4.0",
    "rich>=13.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "jinja2>=3.1",
    "python-multipart>=0.0.9",
    "mcp>=1.27.0",
]

[project.optional-dependencies]
ai = ["anthropic>=0.40"]

[project.scripts]
phone-farm = "phone_farm.cli:cli"
phone-farm-mcp = "phone_farm.mcp_server:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.5",
    "httpx>=0.27",
]
